requirements.txt
streamlit==1.37.0
pandas==2.2.1
openpyxl==3.1.2
Pillow==10.3.0
numpy==1.26.4
statistics
